+++ src/main/java/org/realityforge/guiceyloops/server/glassfish/Provisioner.java
@@ -1,0 +1,9 @@
+package org.realityforge.guiceyloops.server.glassfish;
+
+import javax.annotation.Nonnull;
+
+public interface Provisioner
+{
+  void provision( @Nonnull GlassFishContainer glassfish )
+    throws Exception;
+}